Fix protocol to allow max bound parameters and fields
Bound parameters max out at 65535 due to the FE/BE protocol using a 16bit integer to report the number.
Fields max out at 1664 due to an internal PostgreSQL limit; although it also uses a 16bit integer in the protocol.

+++ src/main/java/com/impossibl/postgres/protocol/ResultField.java
@@ -39,7 +39,7 @@ public class ResultField {
 
   public String name;
   public int relationId;
-  public short relationAttributeNumber;
+  public int relationAttributeNumber;
   public TypeRef typeRef;
   public short typeLength;
   public int typeModifier;
